requirements.txt
Django==4.2.8
djangorestframework==3.14.0
drf-spectacular==0.26.5
stripe==7.7.0
